+++ lib/geolocator.dart
@@ -5,7 +5,9 @@ import 'dart:convert';
 
 import 'package:flutter/services.dart';
 import 'package:meta/meta.dart';
+import 'package:permission_handler/permission_handler.dart';
 
+part 'models/geolocation_enums.dart';
 part 'models/location_accuracy.dart';
 part 'models/location_options.dart';
 part 'models/placemark.dart';
@@ -35,16 +37,40 @@ class Geolocator {
 
   Stream<Position> _onPositionChanged;
 
+  /// Returns a [Future] containing the current [GeolocationStatus] indicating the availability of location services on the device.
+  static Future<GeolocationStatus> checkGeolocationStatus(
+      [GeolocationPermission locationPermission =
+          GeolocationPermission.location]) async {
+    PermissionStatus permissionStatus =
+        await PermissionHandler.checkPermissionStatus(
+            _GeolocationStatusConverter.toPermissionGroup(locationPermission));
+
+    return _GeolocationStatusConverter.fromPermissionStatus(permissionStatus);
+  }
+
   /// Returns the current position taking the supplied [desiredAccuracy] into account.
   ///
   /// When the [desiredAccuracy] is not supplied, it defaults to best.
   Future<Position> getCurrentPosition(
       [LocationAccuracy desiredAccuracy = LocationAccuracy.best]) async {
-    var locationOptions =
-        LocationOptions(accuracy: desiredAccuracy, distanceFilter: 0);
-    var position = await _methodChannel.invokeMethod(
-        'getCurrentPosition', Codec.encodeLocationOptions(locationOptions));
-    return Position._fromMap(position);
+    PermissionStatus permission = await _getLocationPermission();
+
+    if (permission == PermissionStatus.granted) {
+      var locationOptions =
+          LocationOptions(accuracy: desiredAccuracy, distanceFilter: 0);
+      var position = await _methodChannel.invokeMethod(
+          'getCurrentPosition', Codec.encodeLocationOptions(locationOptions));
+
+      try {
+        return Position._fromMap(position);
+      } on ArgumentError {
+        return null;
+      }
+    } else {
+      _handleInvalidPermissions(permission);
+    }
+
+    return null;
   }
 
   /// Returns the last known position stored on the users device.
@@ -54,11 +80,24 @@ class Geolocator {
   /// When no position is available, null is returned.
   Future<Position> getLastKnownPosition(
       [LocationAccuracy desiredAccuracy = LocationAccuracy.best]) async {
-    var locationOptions =
-        LocationOptions(accuracy: desiredAccuracy, distanceFilter: 0);
-    var position = await _methodChannel.invokeMethod(
-        'getLastKnownPosition', Codec.encodeLocationOptions(locationOptions));
-    return Position?._fromMap(position);
+    PermissionStatus permission = await _getLocationPermission();
+
+    if (permission == PermissionStatus.granted) {
+      var locationOptions =
+          LocationOptions(accuracy: desiredAccuracy, distanceFilter: 0);
+      var position = await _methodChannel.invokeMethod(
+          'getLastKnownPosition', Codec.encodeLocationOptions(locationOptions));
+
+      try {
+        return Position._fromMap(position);
+      } on ArgumentError {
+        return null;
+      }
+    } else {
+      _handleInvalidPermissions(permission);
+    }
+
+    return null;
   }
 
   /// Fires whenever the location changes outside the bounds of the [desiredAccuracy].
@@ -80,16 +119,56 @@ class Geolocator {
   /// You can customize the behaviour of the location updates by supplying an
   /// instance [LocationOptions] class. When you don't supply any specific
   /// options, default values will be used for each setting.
-  Stream<Position> getPositionStream(
-      [LocationOptions locationOptions = const LocationOptions()]) {
-    if (_onPositionChanged == null) {
-      _onPositionChanged = _eventChannel
-          .receiveBroadcastStream(Codec.encodeLocationOptions(locationOptions))
-          .map<Position>(
-              (element) => Position._fromMap(element.cast<String, double>()));
+  Future<Stream<Position>> getPositionStream(
+      [LocationOptions locationOptions = const LocationOptions()]) async {
+    PermissionStatus permission = await _getLocationPermission();
+
+    if (permission == PermissionStatus.granted) {
+      if (_onPositionChanged == null) {
+        _onPositionChanged = _eventChannel
+            .receiveBroadcastStream(
+                Codec.encodeLocationOptions(locationOptions))
+            .map<Position>(
+                (element) => Position._fromMap(element.cast<String, double>()));
+      }
+
+      return _onPositionChanged;
+    } else {
+      _handleInvalidPermissions(permission);
     }
 
-    return _onPositionChanged;
+    return null;
+  }
+
+  Future<PermissionStatus> _getLocationPermission() async {
+    PermissionStatus permission =
+        await PermissionHandler.checkPermissionStatus(PermissionGroup.location);
+
+    if (permission != PermissionStatus.granted &&
+        permission != PermissionStatus.disabled) {
+      Map<PermissionGroup, PermissionStatus> permissionStatus =
+          await PermissionHandler
+              .requestPermissions([PermissionGroup.location]);
+
+      return permissionStatus[PermissionGroup.location] ??
+          PermissionStatus.unknown;
+    } else {
+      return permission;
+    }
+  }
+
+  void _handleInvalidPermissions(PermissionStatus permission) {
+    if (permission == PermissionStatus.denied) {
+      throw new PlatformException(
+          code: "PERMISSION_DENIED",
+          message: "Access to location data denied",
+          details: null);
+    } else if (permission == PermissionStatus.disabled) {
+      throw new PlatformException(
+          code: "PERMISSION_DISABLED",
+          message: "Location data is not available on device",
+          details: null);
+    }
   }
 
   /// Returns a list of [Placemark] instances found for the supplied address.
@@ -113,7 +192,12 @@ class Geolocator {
     var placemarks = await _methodChannel.invokeMethod(
         'placemarkFromCoordinates',
         <String, double>{"latitude": latitude, "longitude": longitude});
-    return Placemark._fromMaps(placemarks);
+
+    try {
+      return Placemark._fromMaps(placemarks);
+    } on ArgumentError {
+      return null;
+    }
   }
 
   /// Returns the distance between the supplied coordinates in meters.